Move compare functionality into sub-submodule
