+++ runtime/src/bank/partitioned_epoch_rewards/mod.rs
@@ -1,4 +1,5 @@
 mod calculation;
+mod compare;
 mod distribution;
 mod epoch_rewards_hasher;
 mod sysvar;
